Comprehensive security audit & hardening: eliminate format string vulnerabilities, bounds errors, and pointer-memaccess bugs
- Immunize format specifier calls across qagame, cgame, and ui (G_Printf, G_LogPrintf, CG_Printf, G_Error) by passing explicit '%s' literal format strings.
- Fix double format string evaluation in client chat logging and voice chat notices.
- Add strict bounds validation on client commands, entity slots, weapon indices, and vote targets in g_cmds.c and g_vote.c.
- Sanitize custom shrubbot commands against command chaining and injection (strip ';', '\n', '\r').
- Fix session parser buffer overflow by clamping NQ-GUID reads to '%32s', and eliminate SIGSEGV write into read-only string literal on non-AUTO_GUID builds.
- Fix array overflow in cg_local.h (expand sfx_artilleryDist_1[3]) preventing corruption of sfx_airstrikeExp[0].
- Remediate all 7 legacy sizeof-pointer-memaccess bugs across g_xpsave.c, md5.c, g_character.c, cg_character.c, bg_animation.c, and cg_players.c.
- Passes strict compiler diagnostic scan with 0 warnings across all 120 C source files (-Wformat=2, -Wformat-security, -Wsizeof-pointer-memaccess, -Warray-bounds).

+++ src/cgame/cg_main.c
@@ -113,7 +113,7 @@ Q_EXPORT intptr_t vmMain( intptr_t command, intptr_t arg0, intptr_t arg1, intptr
 */
 		return -1;
 	default:
-		CG_Error( "vmMain: unknown command %i", command );
+		CG_Error( "vmMain: unknown command %i", (int)command );
 		break;
 	}
 	return -1;
